requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.11.0+cu128
torchvision==0.26.0+cu128
numpy
opencv-python
mediapipe==0.10.21
Pillow
onnxruntime-gpu
flask
scikit-image
aiohttp
